Disk optimization

+++ esmecata/annotation.py
@@ -1,4 +1,5 @@
-# Copyright (C) 2021-2023 Arnaud Belcour - Inria, Univ Rennes, CNRS, IRISA Dyliss
+# Copyright (C) 2021-2024 Arnaud Belcour - Inria, Univ Rennes, CNRS, IRISA Dyliss
+# Univ. Grenoble Alpes, Inria, Microcosme
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
@@ -1179,7 +1180,17 @@ def annotate_proteins(input_folder, output_folder, uniprot_sparql_endpoint,
         with open(proteome_tax_id_file, 'r') as tax_id_file:
             csvreader = csv.DictReader(tax_id_file, delimiter='\t')
             for line in csvreader:
-                input_proteomes[line['observation_name']] = line['proteome'].split(',')
+                input_proteomes[line['name']] = line['proteome'].split(',')
+
+    taxon_name_to_observation_name = {}
+    with open(proteome_tax_id_file, 'r') as tax_id_file:
+        csvreader = csv.DictReader(tax_id_file, delimiter='\t')
+        for line in csvreader:
+            taxon_name = line['name'].replace(' ', '_')
+            if taxon_name not in taxon_name_to_observation_name:
+                taxon_name_to_observation_name[taxon_name] = [line['observation_name']]
+            else:
+                taxon_name_to_observation_name[taxon_name].append(line['observation_name'])
 
     reference_protein_path = os.path.join(input_folder, 'reference_proteins')
     # There are 2 ways to handle already annotated proteins:
@@ -1257,9 +1268,6 @@ def annotate_proteins(input_folder, output_folder, uniprot_sparql_endpoint,
         else:
             expression_output_dict = None
         protein_annotations = propagate_annotation_in_cluster(output_dict, reference_proteins, propagate_annotation, uniref_output_dict)
-        annotation_reference_file = os.path.join(annotation_reference_folder, base_filename+'.tsv')
-        write_annotation_reference(protein_annotations, reference_proteins, annotation_reference_file, expression_output_dict)
-        write_pathologic_file(protein_annotations, reference_proteins, pathologic_folder, base_filename, set_proteins)
 
         gos = [go for protein in protein_annotations for go in protein_annotations[protein][1]]
         unique_gos = set(gos)
@@ -1267,6 +1275,11 @@ def annotate_proteins(input_folder, output_folder, uniprot_sparql_endpoint,
         unique_ecs = set(ecs)
         logger.info('|EsMeCaTa|annotation| %d Go Terms (with %d unique GO Terms) and %d EC numbers (with %d unique EC) associated with %s.', len(gos),
                                                                                                 len(unique_gos), len(ecs), len(unique_ecs), base_filename)
+
+        for observation_name in taxon_name_to_observation_name[base_filename]:
+            annotation_reference_file = os.path.join(annotation_reference_folder, observation_name+'.tsv')
+            write_annotation_reference(protein_annotations, reference_proteins, annotation_reference_file, expression_output_dict)
+            write_pathologic_file(protein_annotations, reference_proteins, pathologic_folder, observation_name, set_proteins)
 
     # Create mpwt taxon ID file.
     clustering_taxon_id = {}
@@ -1289,7 +1302,13 @@ def annotate_proteins(input_folder, output_folder, uniprot_sparql_endpoint,
     duration = endtime - starttime
     uniprot_releases['esmecata_annotation_duration'] = duration
     uniprot_metadata_file = os.path.join(output_folder, 'esmecata_metadata_annotation.json')
-    with open(uniprot_metadata_file, 'w') as ouput_file:
-        json.dump(uniprot_releases, ouput_file, indent=4)
+    if os.path.exists(uniprot_metadata_file):
+        metadata_files = [metadata_file for metadata_file in os.listdir(output_folder) if 'esmecata_metadata_annotation' in metadata_file]
+        uniprot_metadata_file = os.path.join(output_folder, 'esmecata_metadata_annotation_{0}.json'.format(len(metadata_files)))
+        with open(uniprot_metadata_file, 'w') as ouput_file:
+            json.dump(uniprot_releases, ouput_file, indent=4)
+    else:
+        with open(uniprot_metadata_file, 'w') as ouput_file:
+            json.dump(uniprot_releases, ouput_file, indent=4)
 
-    logger.info('|EsMeCaTa|annotation| Annotation complete.')
+    logger.info('|EsMeCaTa|annotation| Annotation complete in {0}.'.format(duration))
